# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing delivery agents, orders and allocation system",
        "title": "Delivery Management System API",
        "contact": {
            "name": "Prince Raj",
            "email": "prince@example.com"
        },
        "version": "1.0"
    },
    "host": "localhost:5002",
    "basePath": "/",
    "paths": {
        "/api/agent-summary": {
            "get": {
                "description": "Returns a paginated summary of agents, including total orders, distance, time, and profit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Summary"
                ],
                "summary": "Get agent summary with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default is 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.PaginatedAgentSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/agent/checkin": {
            "post": {
                "description": "Allows an agent to check in to a warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "Check-in an agent",
                "parameters": [
                    {
                        "description": "Agent Check-In Info",
                        "name": "agent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AgentCheckInRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/allocate": {
            "get": {
                "description": "Runs the allocation algorithm to assign orders to agents",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Trigger manual allocation of orders",
                "responses": {
                    "200": {
                        "description": "Allocation successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Allocation failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/assignments": {
            "get": {
                "description": "Returns paginated list of assignments with formatted date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Assignments"
                ],
                "summary": "Get paginated assignments",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of assignments",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/checkin": {
            "get": {
                "description": "Marks the agent as checked-in at the warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent"
                ],
                "summary": "Check in an agent (second version)",
                "parameters": [
                    {
                        "description": "Agent Check-In Data",
                        "name": "agent",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.AgentCheckInRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/order": {
            "post": {
                "description": "Creates a new customer order and stores it in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order details",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/orders/bulk": {
            "post": {
                "description": "Accepts a list of customer orders and stores them in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Create multiple orders in bulk",
                "parameters": [
                    {
                        "description": "List of order requests",
                        "name": "orders",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.OrderRequest"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/system-summary": {
            "get": {
                "description": "Returns a system-wide summary including total, assigned, and deferred orders, along with agent utilization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Summary"
                ],
                "summary": "Get system summary with paginated agent utilization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default is 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.SystemSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/warehouse": {
            "post": {
                "description": "Accepts warehouse details and stores them in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Warehouse"
                ],
                "summary": "Create a new warehouse",
                "parameters": [
                    {
                        "description": "Warehouse Details",
                        "name": "warehouse",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.WarehouseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "types.AgentCheckInRequest": {
            "type": "object",
            "required": [
                "name",
                "warehouse_id"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
        "types.AgentSummary": {
            "type": "object",
            "properties": {
                "agent_id": {
                    "type": "integer"
                },
                "profit": {
                    "type": "number"
                },
                "total_km": {
                    "type": "number"
                },
                "total_minutes": {
                    "type": "number"
                },
                "total_orders": {
                    "type": "integer"
                }
            }
        },
        "types.Location": {
            "type": "object",
            "required": [
                "lat",
                "lng"
            ],
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
        "types.OrderRequest": {
            "type": "object",
            "required": [
                "customer",
                "lat",
                "lng",
                "warehouse_id"
            ],
            "properties": {
                "customer": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
        "types.PaginatedAgentSummary": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.AgentSummary"
                    }
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "types.SystemSummary": {
            "type": "object",
            "properties": {
                "agent_utilization": {
                    "$ref": "#/definitions/types.PaginatedAgentSummary"
                },
                "assigned_orders": {
                    "type": "integer"
                },
                "deferred_orders": {
                    "type": "integer"
                },
                "total_orders": {
                    "type": "integer"
                }
            }
        },
        "types.WarehouseRequest": {
            "type": "object",
            "required": [
                "location",
                "name"
            ],
            "properties": {
                "location": {
                    "$ref": "#/definitions/types.Location"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}